1. refactored CIM with various algos; 2. lint
