+++ examples/rl/cim/__init__.py
@@ -3,6 +3,6 @@
 
 from .callbacks import post_collect, post_evaluate
 from .env_sampler import agent2policy, get_env_sampler
-from .policies import policy_creator, trainer_creator
+from .policy_trainer import policy_creator, trainer_creator
 
 __all__ = ["agent2policy", "post_collect", "post_evaluate", "get_env_sampler", "policy_creator", "trainer_creator"]
