+++ examples/rl/cim/algorithms/ac.py
@@ -1,35 +1,24 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
 from typing import Dict
 
 import torch
 from torch.optim import Adam, RMSprop
 
-from maro.rl_v3.model import DiscretePolicyNet, DiscreteQNet, FullyConnected, VNet
+from maro.rl_v3.model import DiscretePolicyNet, FullyConnected, VNet
+from maro.rl_v3.policy import DiscretePolicyGradient
+from maro.rl_v3.training.algorithms import DiscreteActorCritic, DiscreteActorCriticParams
 
-from .config import action_shaping_conf, state_dim
 
-q_net_conf = {
-    "input_dim": state_dim,
-    "hidden_dims": [256, 128, 64, 32],
-    "output_dim": len(action_shaping_conf["action_space"]),
-    "activation": torch.nn.LeakyReLU,
-    "softmax": False,
-    "batch_norm": True,
-    "skip_connection": False,
-    "head": True,
-    "dropout_p": 0.0
-}
-q_net_optim_conf = (RMSprop, {"lr": 0.05})
 actor_net_conf = {
-    "input_dim": state_dim,
     "hidden_dims": [256, 128, 64],
-    "output_dim": len(action_shaping_conf["action_space"]),
     "activation": torch.nn.Tanh,
     "softmax": True,
     "batch_norm": False,
     "head": True
 }
 critic_net_conf = {
-    "input_dim": state_dim,
     "hidden_dims": [256, 128, 64],
     "output_dim": 1,
     "activation": torch.nn.LeakyReLU,
@@ -41,45 +30,10 @@ actor_optim_conf = (Adam, {"lr": 0.001})
 critic_optim_conf = (RMSprop, {"lr": 0.001})
 
 
-# #####################################################################################################################
-class MyQNet(DiscreteQNet):
-    def __init__(self) -> None:
-        super(MyQNet, self).__init__(state_dim=q_net_conf["input_dim"], action_num=q_net_conf["output_dim"])
-        self._fc = FullyConnected(**q_net_conf)
-        self._optim = q_net_optim_conf[0](self._fc.parameters(), **q_net_optim_conf[1])
-
-    def _get_q_values_for_all_actions(self, states: torch.Tensor) -> torch.Tensor:
-        return self._fc(states)
-
-    def get_gradients(self, loss: torch.Tensor) -> Dict[str, torch.Tensor]:
-        self._optim.zero_grad()
-        loss.backward()
-        return {name: param.grad for name, param in self.named_parameters()}
-
-    def apply_gradients(self, grad: dict) -> None:
-        for name, param in self.named_parameters():
-            param.grad = grad[name]
-        self._optim.step()
-
-    def get_net_state(self) -> object:
-        return {"network": self.state_dict(), "optim": self._optim.state_dict()}
-
-    def set_net_state(self, net_state: object) -> None:
-        assert isinstance(net_state, dict)
-        self.load_state_dict(net_state["network"])
-        self._optim.load_state_dict(net_state["optim"])
-
-    def freeze(self) -> None:
-        self.freeze_all_parameters()
-
-    def unfreeze(self) -> None:
-        self.unfreeze_all_parameters()
-
-
 class MyActorNet(DiscretePolicyNet):
-    def __init__(self) -> None:
-        super(MyActorNet, self).__init__(state_dim=actor_net_conf["input_dim"], action_num=actor_net_conf["output_dim"])
-        self._actor = FullyConnected(**actor_net_conf)
+    def __init__(self, state_dim: int, action_num: int) -> None:
+        super(MyActorNet, self).__init__(state_dim=state_dim, action_num=action_num)
+        self._actor = FullyConnected(input_dim=state_dim, output_dim=action_num, **actor_net_conf)
         self._actor_optim = actor_optim_conf[0](self._actor.parameters(), **actor_optim_conf[1])
 
     def _get_action_probs_impl(self, states: torch.Tensor) -> torch.Tensor:
@@ -113,9 +67,9 @@ class MyActorNet(DiscretePolicyNet):
 
 
 class MyCriticNet(VNet):
-    def __init__(self) -> None:
-        super(MyCriticNet, self).__init__(state_dim=critic_net_conf["input_dim"])
-        self._critic = FullyConnected(**critic_net_conf)
+    def __init__(self, state_dim: int) -> None:
+        super(MyCriticNet, self).__init__(state_dim=state_dim)
+        self._critic = FullyConnected(input_dim=state_dim, **critic_net_conf)
         self._critic_optim = critic_optim_conf[0](self._critic.parameters(), **critic_optim_conf[1])
 
     def _get_v_values(self, states: torch.Tensor) -> torch.Tensor:
@@ -146,3 +100,24 @@ class MyCriticNet(VNet):
 
     def unfreeze(self) -> None:
         self.unfreeze_all_parameters()
+
+
+def get_discrete_policy_gradient(name: str, *, state_dim: int, action_num: int) -> DiscretePolicyGradient:
+    return DiscretePolicyGradient(name=name, policy_net=MyActorNet(state_dim, action_num))
+
+
+def get_ac(name: str, *, state_dim: int) -> DiscreteActorCritic:
+    return DiscreteActorCritic(
+        name=name,
+        params=DiscreteActorCriticParams(
+            device="cpu",
+            get_v_critic_net_func=lambda: MyCriticNet(state_dim),
+            reward_discount=.0,
+            grad_iters=10,
+            critic_loss_cls=torch.nn.SmoothL1Loss,
+            min_logp=None,
+            lam=.0,
+        )
+    )
+
+
