+++ examples/rl/cim/algorithms/maddpg.py
@@ -1,27 +1,25 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
+from functools import partial
 from typing import Dict, List
 
 import torch
 from torch.optim import Adam, RMSprop
 
 from maro.rl_v3.model import DiscretePolicyNet, FullyConnected, MultiQNet
+from maro.rl_v3.policy import DiscretePolicyGradient
+from maro.rl_v3.training.algorithms import DiscreteMADDPG, DiscreteMADDPGParams
 
-from .config import action_shaping_conf, state_dim
 
 actor_net_conf = {
-    "input_dim": state_dim,
     "hidden_dims": [256, 128, 64],
-    "output_dim": len(action_shaping_conf["action_space"]),
     "activation": torch.nn.Tanh,
     "softmax": True,
     "batch_norm": False,
     "head": True
 }
-critic_conf = {
-    "state_dim": state_dim,
-    "action_dims": [1]
-}
 critic_net_conf = {
-    "input_dim": critic_conf["state_dim"] + sum(critic_conf["action_dims"]),
     "hidden_dims": [256, 128, 64],
     "output_dim": 1,
     "activation": torch.nn.LeakyReLU,
@@ -35,9 +33,9 @@ critic_optim_conf = (RMSprop, {"lr": 0.001})
 
 # #####################################################################################################################
 class MyActorNet(DiscretePolicyNet):
-    def __init__(self) -> None:
-        super(MyActorNet, self).__init__(state_dim=actor_net_conf["input_dim"], action_num=actor_net_conf["output_dim"])
-        self._actor = FullyConnected(**actor_net_conf)
+    def __init__(self, state_dim: int, action_num: int) -> None:
+        super(MyActorNet, self).__init__(state_dim=state_dim, action_num=action_num)
+        self._actor = FullyConnected(input_dim=state_dim, output_dim=action_num, **actor_net_conf)
         self._actor_optim = actor_optim_conf[0](self._actor.parameters(), **actor_optim_conf[1])
 
     def _get_action_probs_impl(self, states: torch.Tensor) -> torch.Tensor:
@@ -71,12 +69,9 @@ class MyActorNet(DiscretePolicyNet):
 
 
 class MyMultiCriticNet(MultiQNet):
-    def __init__(self) -> None:
-        super(MyMultiCriticNet, self).__init__(
-            state_dim=critic_conf["state_dim"],
-            action_dims=critic_conf["action_dims"]
-        )
-        self._critic = FullyConnected(**critic_net_conf)
+    def __init__(self, state_dim: int, action_dims: List[int]) -> None:
+        super(MyMultiCriticNet, self).__init__(state_dim=state_dim, action_dims=action_dims)
+        self._critic = FullyConnected(input_dim=state_dim + sum(action_dims), **critic_net_conf)
         self._critic_optim = critic_optim_conf[0](self._critic.parameters(), **critic_optim_conf[1])
 
     def _get_q_values(self, states: torch.Tensor, actions: List[torch.Tensor]) -> torch.Tensor:
@@ -107,3 +102,24 @@ class MyMultiCriticNet(MultiQNet):
 
     def unfreeze(self) -> None:
         self.unfreeze_all_parameters()
+
+
+def get_multi_critic_net(state_dim: int, action_dims: List[int]) -> MyMultiCriticNet:
+    return MyMultiCriticNet(state_dim, action_dims)
+
+
+def get_discrete_policy_gradient(name: str, *, state_dim: int, action_num: int) -> DiscretePolicyGradient:
+    return DiscretePolicyGradient(name=name, policy_net=MyActorNet(state_dim, action_num))
+
+
+def get_maddpg(name: str, *, state_dim: int, action_dims: List[int]) -> DiscreteMADDPG:
+    return DiscreteMADDPG(
+        name=name,
+        params=DiscreteMADDPGParams(
+            device="cpu",
+            reward_discount=.0,
+            num_epoch=10,
+            get_q_critic_net_func=partial(get_multi_critic_net, state_dim, action_dims),
+            # shared_critic=True,
+        )
+    )
