+++ maro/cli/utils/docker.py
@@ -1,9 +1,6 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-import os
-import subprocess
-
 import docker
