+++ maro/rl_v3/distributed/__init__.py
@@ -1,5 +1,8 @@
-from .remote_obj import Client, RemoteObj
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
 from .dispatcher import Dispatcher
+from .remote_obj import Client, RemoteObj
 from .worker import Worker
 
-__all__ = ["Dispatcher", "Worker"]
+__all__ = ["Client", "Dispatcher", "RemoteObj", "Worker"]
